fix prometheus dollar escape
