+++ translator/tocwconfig/sampleConfig/opentelemetry/combined_v1_v2_eks_config.yaml
@@ -587,9 +587,9 @@ processors:
         ec2_instance_tag_keys:
             - AutoScalingGroupName
         ec2_metadata_tags:
-            - InstanceType
             - ImageId
             - InstanceId
+            - InstanceType
         imds_retries: 1
         middleware: agenthealth/statuscode
         refresh_tags_interval: 0s
@@ -2084,7 +2084,7 @@ receivers:
                   relabel_configs:
                     - action: replace
                       regex: (.*)
-                      replacement: $$1
+                      replacement: $$$$1
                       separator: ;
                       source_labels:
                         - __meta_kubernetes_pod_name
@@ -2426,13 +2426,13 @@ service:
                 - ec2tagger
                 - awsentity/resource
             receivers:
-                - telegraf_procstat/1917393364
-                - telegraf_processes
                 - telegraf_disk
-                - telegraf_mem
-                - telegraf_cpu
                 - telegraf_swap
+                - telegraf_processes
+                - telegraf_cpu
+                - telegraf_mem
                 - telegraf_netstat
+                - telegraf_procstat/1917393364
         metrics/host_metrics:
             exporters:
                 - forward/opentelemetry
@@ -2447,8 +2447,8 @@ service:
                 - ec2tagger
                 - awsentity/service/telegraf
             receivers:
-                - telegraf_statsd
                 - telegraf_socket_listener
+                - telegraf_statsd
         metrics/hostDeltaMetrics:
             exporters:
                 - awscloudwatch
@@ -2457,8 +2457,8 @@ service:
                 - ec2tagger
                 - awsentity/resource
             receivers:
-                - telegraf_diskio
                 - telegraf_net
+                - telegraf_diskio
         metrics/opentelemetry:
             exporters:
                 - otlphttp/metrics
