+++ translator/tocwconfig/sampleConfig/opentelemetry/prometheus_otel_pipeline_eks_config.yaml
@@ -555,7 +555,7 @@ receivers:
                   relabel_configs:
                     - action: replace
                       regex: (.*)
-                      replacement: $$1
+                      replacement: $$$$1
                       separator: ;
                       source_labels:
                         - __meta_kubernetes_pod_name
