+++ translator/translate/otel/pipeline/opentelemetry/prometheus/translator_test.go
@@ -163,6 +163,31 @@ func TestPrometheusReceiverTranslator(t *testing.T) {
 	assert.NotNil(t, cfg)
 }
 
+func TestEscapeDollarDigit(t *testing.T) {
+	tests := []struct {
+		name string
+		in   string
+		want string
+	}{
+		{name: "bare $1", in: "replacement: $1", want: "replacement: $$$$1"},
+		{name: "bare $0", in: "$0", want: "$$$$0"},
+		{name: "bare $9", in: "value=$9", want: "value=$$$$9"},
+		{name: "multiple refs", in: "$1 and $2 and $3", want: "$$$$1 and $$$$2 and $$$$3"},
+		{name: "dollar non-digit", in: "$HOME and $PATH", want: "$HOME and $PATH"},
+		{name: "no dollar", in: "no dollars here", want: "no dollars here"},
+		{name: "empty string", in: "", want: ""},
+		{name: "dollar at end", in: "trailing$", want: "trailing$"},
+		{name: "multi-digit $10", in: "$10", want: "$$$$10"},
+		{name: "mixed text", in: "tag: k8s.label.$1 and $FOO", want: "tag: k8s.label.$$$$1 and $FOO"},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := escapeDollarDigit(tt.in)
+			assert.Equal(t, tt.want, got)
+		})
+	}
+}
+
 func TestPrometheusReceiverTranslatorDollarEscape(t *testing.T) {
 	// Create a prometheus config with relabel_configs that use $1 capture group references.
 	// Without escaping, the OTel expandconverter would treat $1 as an env var and blank it out.
@@ -201,10 +226,10 @@ func TestPrometheusReceiverTranslatorDollarEscape(t *testing.T) {
 
 	relabelCfgs := promCfg.PrometheusConfig.ScrapeConfigs[0].RelabelConfigs
 	require.Len(t, relabelCfgs, 2)
-	// $$1 is the correct value at this stage — the expandconverter resolves $$1 → $1 at
-	// collector startup. Without the escape, $1 would be treated as an env var and blanked out.
-	assert.Equal(t, "$$1", relabelCfgs[0].Replacement, "first relabel $1 should be escaped to $$1 for expandconverter")
-	assert.Equal(t, "$$1:$$2", relabelCfgs[1].Replacement, "second relabel $1:$2 should be escaped to $$1:$$2 for expandconverter")
+	// $$$$1 survives the resolver's escapeDollarSigns ($$$$→$$) and the expandconverter ($$→$)
+	// to produce the final $1 at runtime.
+	assert.Equal(t, "$$$$1", relabelCfgs[0].Replacement, "first relabel $1 should be escaped to $$$$1 for resolver+expandconverter")
+	assert.Equal(t, "$$$$1:$$$$2", relabelCfgs[1].Replacement, "second relabel $1:$2 should be escaped to $$$$1:$$$$2 for resolver+expandconverter")
 }
 
 func TestPrometheusReceiverTranslatorMissingFile(t *testing.T) {
